Move more of the query DSL implementation (result definition) to the engine
