+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/index/impl/ElasticsearchIndexSearchTarget.java
@@ -12,46 +12,30 @@ import java.util.stream.Collectors;
 
 import org.hibernate.search.v6poc.backend.elasticsearch.document.model.impl.ElasticsearchIndexModel;
 import org.hibernate.search.v6poc.backend.elasticsearch.impl.ElasticsearchBackend;
-import org.hibernate.search.v6poc.backend.elasticsearch.logging.impl.Log;
-import org.hibernate.search.v6poc.backend.elasticsearch.orchestration.impl.ElasticsearchWorkOrchestrator;
-import org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl.ElasticsearchSearchPredicateCollector;
-import org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl.ElasticsearchSearchTargetContext;
-import org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl.ElasticsearchSingleSearchPredicateCollector;
-import org.hibernate.search.v6poc.search.dsl.spi.SearchPredicateContributor;
-import org.hibernate.search.v6poc.backend.elasticsearch.search.impl.SearchQueryResultDefinitionContextImpl;
-import org.hibernate.search.v6poc.search.dsl.spi.SearchTargetPredicateRootContext;
-import org.hibernate.search.v6poc.search.predicate.spi.SearchPredicateFactory;
-import org.hibernate.search.v6poc.backend.elasticsearch.search.predicate.impl.SearchPredicateFactoryImpl;
-import org.hibernate.search.v6poc.backend.elasticsearch.work.impl.ElasticsearchWorkFactory;
+import org.hibernate.search.v6poc.backend.elasticsearch.search.query.impl.ElasticsearchSearchTargetContextImpl;
 import org.hibernate.search.v6poc.backend.index.spi.IndexSearchTarget;
 import org.hibernate.search.v6poc.engine.spi.SessionContext;
 import org.hibernate.search.v6poc.search.DocumentReference;
 import org.hibernate.search.v6poc.search.ObjectLoader;
 import org.hibernate.search.v6poc.search.SearchPredicate;
 import org.hibernate.search.v6poc.search.dsl.predicate.SearchPredicateContainerContext;
-import org.hibernate.search.v6poc.search.spi.SearchQueryResultDefinitionContext;
-import org.hibernate.search.v6poc.util.spi.LoggerFactory;
+import org.hibernate.search.v6poc.search.dsl.spi.SearchTargetContext;
+import org.hibernate.search.v6poc.search.dsl.spi.SearchTargetPredicateRootContext;
+import org.hibernate.search.v6poc.search.dsl.query.SearchQueryResultDefinitionContext;
+import org.hibernate.search.v6poc.search.dsl.spi.SearchQueryResultDefinitionContextImpl;
 
 
 /**
  * @author Yoann Rodiere
  */
-class ElasticsearchIndexSearchTarget implements IndexSearchTarget, ElasticsearchSearchTargetContext {
+class ElasticsearchIndexSearchTarget implements IndexSearchTarget {
 
-	private static final Log log = LoggerFactory.make( Log.class );
-
-	private final ElasticsearchBackend backend;
-
-	private final SearchPredicateFactory<ElasticsearchSearchPredicateCollector> searchPredicateFactory;
-
-	private final Set<ElasticsearchIndexModel> indexModels;
+	private final SearchTargetContext<?> searchTargetContext;
 
 	private final Set<String> indexNames;
 
 	ElasticsearchIndexSearchTarget(ElasticsearchBackend backend, Set<ElasticsearchIndexModel> indexModels) {
-		this.backend = backend;
-		this.searchPredicateFactory = new SearchPredicateFactoryImpl( indexModels );
-		this.indexModels = indexModels;
+		this.searchTargetContext = new ElasticsearchSearchTargetContextImpl( backend, indexModels );
 		this.indexNames = indexModels.stream()
 				.map( ElasticsearchIndexModel::getIndexName )
 				.collect( Collectors.toSet() );
@@ -62,61 +46,20 @@ class ElasticsearchIndexSearchTarget implements IndexSearchTarget, Elasticsearch
 			SessionContext context,
 			Function<DocumentReference, R> documentReferenceTransformer,
 			ObjectLoader<R, O> objectLoader) {
-		return new SearchQueryResultDefinitionContextImpl<>( this, context,
+		return new SearchQueryResultDefinitionContextImpl<>( searchTargetContext, context,
 				documentReferenceTransformer, objectLoader );
 	}
 
 	@Override
 	public SearchPredicateContainerContext<SearchPredicate> predicate() {
-		return new SearchTargetPredicateRootContext<>( this );
-	}
-
-	@Override
-	public Set<String> getIndexNames() {
-		return indexNames;
-	}
-
-	@Override
-	public Set<ElasticsearchIndexModel> getIndexModels() {
-		return indexModels;
-	}
-
-	@Override
-	public SearchPredicate toSearchPredicate(SearchPredicateContributor<ElasticsearchSearchPredicateCollector> contributor) {
-		ElasticsearchSingleSearchPredicateCollector collector = new ElasticsearchSingleSearchPredicateCollector();
-		contributor.contribute( collector );
-		return new ElasticsearchSearchPredicate( collector.toJson() );
-	}
-
-	@Override
-	public SearchPredicateContributor<ElasticsearchSearchPredicateCollector> toContributor(SearchPredicate predicate) {
-		if ( !( predicate instanceof ElasticsearchSearchPredicate ) ) {
-			throw log.cannotMixElasticsearchSearchQueryWithOtherPredicates( predicate );
-		}
-		return (ElasticsearchSearchPredicate) predicate;
-	}
-
-	@Override
-	public SearchPredicateFactory<ElasticsearchSearchPredicateCollector> getSearchPredicateFactory() {
-		return searchPredicateFactory;
-	}
-
-	@Override
-	public ElasticsearchWorkFactory getWorkFactory() {
-		return backend.getWorkFactory();
-	}
-
-	@Override
-	public ElasticsearchWorkOrchestrator getQueryOrchestrator() {
-		return backend.getQueryOrchestrator();
+		return new SearchTargetPredicateRootContext<>( searchTargetContext );
 	}
 
 	@Override
 	public String toString() {
 		return new StringBuilder( getClass().getSimpleName() )
 				.append( "[" )
-				.append( "backend=" ).append( backend )
-				.append( ", indexNames=" ).append( indexNames )
+				.append( "indexNames=" ).append( indexNames )
 				.append( "]")
 				.toString();
 	}
