+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/search/dsl/impl/ElasticsearchSearchPredicateContainerContextImpl.java
@@ -7,6 +7,7 @@
 package org.hibernate.search.v6poc.backend.elasticsearch.search.dsl.impl;
 
 import org.hibernate.search.v6poc.backend.elasticsearch.search.ElasticsearchSearchPredicateContainerContext;
+import org.hibernate.search.v6poc.backend.elasticsearch.search.query.impl.ElasticsearchSearchTargetContext;
 import org.hibernate.search.v6poc.search.dsl.predicate.SearchPredicateContainerContext;
 import org.hibernate.search.v6poc.search.dsl.spi.DelegatingSearchPredicateContainerContextImpl;
 import org.hibernate.search.v6poc.search.dsl.spi.SearchDslContext;
