+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/search/query/impl/SearchResultExtractorImpl.java
@@ -4,14 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.backend.elasticsearch.search.impl;
+package org.hibernate.search.v6poc.backend.elasticsearch.search.query.impl;
 
 import java.util.Collections;
 import java.util.List;
 
 import org.hibernate.search.v6poc.backend.elasticsearch.gson.impl.JsonAccessor;
 import org.hibernate.search.v6poc.search.SearchResult;
-import org.hibernate.search.v6poc.search.spi.HitAggregator;
+import org.hibernate.search.v6poc.search.query.spi.HitAggregator;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
