+++ engine/src/main/java/org/hibernate/search/v6poc/search/dsl/query/SearchQueryWrappingDefinitionResultContext.java
@@ -4,11 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.search.spi;
+package org.hibernate.search.v6poc.search.dsl.query;
 
 import java.util.function.Function;
-
-import org.hibernate.search.v6poc.search.dsl.query.SearchQueryResultContext;
 
 /**
  * The context used when building a query, when a query wrapper can be defined.
