+++ engine/src/main/java/org/hibernate/search/v6poc/search/dsl/query/impl/SearchQueryWrappingDefinitionResultContextImpl.java
@@ -4,19 +4,21 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.search.dsl.spi;
+package org.hibernate.search.v6poc.search.dsl.query.impl;
 
 import java.util.function.Consumer;
 import java.util.function.Function;
 
 import org.hibernate.search.v6poc.search.SearchPredicate;
 import org.hibernate.search.v6poc.search.SearchQuery;
+import org.hibernate.search.v6poc.search.dsl.impl.QuerySearchPredicateBuildingRootContextImpl;
 import org.hibernate.search.v6poc.search.dsl.predicate.impl.SearchPredicateContainerContextImpl;
-import org.hibernate.search.v6poc.search.dsl.query.impl.SearchQueryContextImpl;
+import org.hibernate.search.v6poc.search.dsl.query.SearchQueryWrappingDefinitionResultContext;
 import org.hibernate.search.v6poc.search.dsl.predicate.SearchPredicateContainerContext;
 import org.hibernate.search.v6poc.search.dsl.query.SearchQueryContext;
 import org.hibernate.search.v6poc.search.dsl.query.SearchQueryResultContext;
-import org.hibernate.search.v6poc.search.spi.SearchQueryWrappingDefinitionResultContext;
+import org.hibernate.search.v6poc.search.dsl.spi.SearchTargetContext;
+import org.hibernate.search.v6poc.search.query.spi.SearchQueryBuilder;
 
 
 public final class SearchQueryWrappingDefinitionResultContextImpl<T, C, Q>
