+++ engine/src/main/java/org/hibernate/search/v6poc/search/query/impl/SimpleHitAggregator.java
@@ -4,51 +4,55 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.v6poc.search.spi;
+package org.hibernate.search.v6poc.search.query.impl;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Function;
 
-import org.hibernate.search.v6poc.search.ObjectLoader;
+import org.hibernate.search.v6poc.search.query.spi.HitAggregator;
+import org.hibernate.search.v6poc.search.query.spi.HitCollector;
 import org.hibernate.search.v6poc.util.AssertionFailure;
 
-public final class ObjectHitAggregator<R, O> implements HitAggregator<LoadingHitCollector<R>, List<O>> {
+public final class SimpleHitAggregator<T, U> implements HitAggregator<HitCollector<T>, List<U>> {
 
-	private final ObjectLoader<R, O> objectLoader;
+	private final Function<T, U> hitTransformer;
+
 	private final HitCollectorImpl hitCollector = new HitCollectorImpl();
 
-	private final ArrayList<R> referencesToLoad = new ArrayList<>();
+	private List<U> hits;
 
-	public ObjectHitAggregator(ObjectLoader<R, O> objectLoader) {
-		this.objectLoader = objectLoader;
+	public SimpleHitAggregator(Function<T, U> hitTransformer) {
+		this.hitTransformer = hitTransformer;
 	}
 
 	@Override
 	public void init(int expectedHitCount) {
-		referencesToLoad.clear();
-		referencesToLoad.ensureCapacity( expectedHitCount );
+		hits = new ArrayList<>( expectedHitCount );
 	}
 
 	@Override
-	public LoadingHitCollector<R> nextCollector() {
+	public HitCollector<T> nextCollector() {
 		hitCollector.reset();
 		return hitCollector;
 	}
 
 	@Override
-	public List<O> build() {
-		return objectLoader.load( referencesToLoad );
+	public List<U> build() {
+		List<U> result = hits;
+		hits = null;
+		return result;
 	}
 
-	private class HitCollectorImpl implements LoadingHitCollector<R> {
+	private class HitCollectorImpl implements HitCollector<T> {
 
 		private boolean currentHitCollected = false;
 
 		@Override
-		public void collectForLoading(R reference) {
+		public void collect(T hit) {
 			checkNotAlreadyCollected();
 			currentHitCollected = true;
-			referencesToLoad.add( reference );
+			hits.add( hitTransformer.apply( hit ) );
 		}
 
 		public void reset() {
