+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/mapping/PojoSearchTarget.java
@@ -9,7 +9,7 @@ package org.hibernate.search.v6poc.entity.pojo.mapping;
 import org.hibernate.search.v6poc.entity.pojo.search.PojoReference;
 import org.hibernate.search.v6poc.search.SearchPredicate;
 import org.hibernate.search.v6poc.search.dsl.predicate.SearchPredicateContainerContext;
-import org.hibernate.search.v6poc.search.spi.SearchQueryResultDefinitionContext;
+import org.hibernate.search.v6poc.search.dsl.query.SearchQueryResultDefinitionContext;
 
 public interface PojoSearchTarget<T> {
